Add loggoutOn404 flag
The ARequest service has a flag on it's methods now, that can optionally be passed
in by the caller to decide whether the user should be redirected to the login page
or whether fellow players failed in their log in

+++ frontend/static/js/services/ARequestService.js
@@ -10,26 +10,17 @@ class ARequestService {
 		}
 	}
 
-	throwCustomError(errorCode, errorText) {
-		const errorMap = {
-			401: customErrors.unauthorized_401,
-			409: customErrors.conflict_409,
-		};
 
-		const errorClass = errorMap[errorCode] || Error;
-		throw new errorClass(errorText);
-	}
-
-	async checkResponseWithBody(request) {
+	async checkResponseWithBody(request, logoutOn401 = true) {
 		try {
 			const response = await request;
 			if (!response.ok) {
-				if (response.status === 401) {
+				if (response.status === 401 && logoutOn401) {
 					localStorage.clear();
 					navigateTo('/login');
 					return;
 				}
-				this.throwCustomError(response.status, response.statusText);
+				throw new Error(response.message);
 			}
 			return response.json();
 		} catch (error) {
@@ -41,10 +32,15 @@ class ARequestService {
 		}
 	}
 
-	async checkResponseNoBody(request) {
+	async checkResponseNoBody(request, logoutOn401 = true) {
 		try {
 			const response = await request;
 			if (!response.ok) {
+				if (response.status === 401) {
+					localStorage.clear();
+					navigateTo('/login');
+					return;
+				}
 				throw new Error('Error: ' + response.status);
 			}
 			return '';
@@ -57,21 +53,21 @@ class ARequestService {
 		}
 	}
 
-	async getRequest(url) {
+	async getRequest(url, logoutOn401 = true) {
 		const request = fetch(`${url}`, {
 			credentials: 'include',
 		});
-		return this.checkResponseWithBody(request);
+		return this.checkResponseWithBody(request, logoutOn401);
 	}
 
-	async getAllRequest(url) {
+	async getAllRequest(url, logoutOn401 = true) {
 		const request = fetch(`${url}`, {
 			credentials: 'include',
 		});
-		return this.checkResponseWithBody(request);
+		return this.checkResponseWithBody(request, logoutOn401);
 	}
 
-	async postRequest(url, jsonBody) {
+	async postRequest(url, jsonBody, logoutOn401 = true) {
 		const request = fetch(`${url}`, {
 			method: 'POST',
 			headers: {
@@ -82,10 +78,10 @@ class ARequestService {
 			credentials: 'include',
 		});
 
-		return this.checkResponseWithBody(request);
+		return this.checkResponseWithBody(request, logoutOn401);
 	}
 
-	async putRequest(url, id, jsonBody) {
+	async putRequest(url, id, jsonBody, logoutOn401 = true) {
 		const request = fetch(`${url}${id}`, {
 			method: 'PUT',
 			headers: {
@@ -96,10 +92,10 @@ class ARequestService {
 			credentials: 'include',
 		});
 
-		return this.checkResponseWithBody(request);
+		return this.checkResponseWithBody(request, logoutOn401);
 	}
 
-	async deleteRequest(url) {
+	async deleteRequest(url, logoutOn401 = true) {
 		const request = fetch(`${url}`, {
 			method: 'DELETE',
 			headers: {
@@ -109,7 +105,7 @@ class ARequestService {
 			credentials: 'include',
 		});
 
-		return this.checkResponseNoBody(request);
+		return this.checkResponseNoBody(request, logoutOn401);
 	}
 }
